server/requirements.txt
flask
google-cloud-firestore